requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.3
Pillow==11.2.1
SpeechRecognition==3.10.0
pydub==0.25.1
numpy==1.25.2
matplotlib==3.7.2
requests==2.31.0
